Refactor task management components and update task interface
