+++ frontend/app/features/kanban/components/KanbanBoard.tsx
@@ -4,6 +4,9 @@ import { useGetAllTasksQuery } from "@store/api/taskSlice";
 import { useState } from "react";
 import { useTaskFilters } from "../hooks/useTaskFilters";
 import styles from "../styles/kanban.module.css";
+import CreateTaskForm from "./CreateTaskForm";
+import EditTaskForm from "./EditTaskForm";
+import TaskCard from "./TaskCard";
 import TaskFilter from "./TaskFilter";
 
 const KanbanBoard: React.FC = () => {
@@ -14,8 +17,14 @@ const KanbanBoard: React.FC = () => {
   const { data: tasksData, error, isLoading } = useGetAllTasksQuery(filters);
   const tasks = Array.isArray(tasksData) ? tasksData : [];
 
+  const [selectedTask, setSelectedTask] = useState(null);
+  const [isEditOpen, setIsEditOpen] = useState(false);
+
   if (isLoading) return <p className="notification is-info">Loading tasks...</p>;
-  if (error) return <p className="notification is-danger">Error loading tasks.</p>;
+  if (error) {
+    console.error("Error loading tasks:", error);
+    return <p className="notification is-danger">Error loading tasks.</p>;
+  }
 
   return (
     <div className={styles.kanbanContainer}>
@@ -31,6 +40,8 @@ const KanbanBoard: React.FC = () => {
         uniqueAssignees={uniqueAssignees}
       />
 
+      <CreateTaskForm />
+
       <div className={`${styles.kanbanBoard} columns is-variable is-4`}>
         {["TODO", "IN_PROGRESS", "DONE"].map((status) => (
           <div key={status} className="column is-one-third">
@@ -42,23 +53,19 @@ const KanbanBoard: React.FC = () => {
                 tasks
                   .filter((task) => task.status === status)
                   .map((task) => (
-                    <div key={task.id} className="card mb-3">
-                      <header className="card-header">
-                        <p className="card-header-title">{task.title}</p>
-                      </header>
-                      <div className="card-content">
-                        <p>{task.description}</p>
-                        <p className="has-text-grey">
-                          Created By: {task.createdBy} | Assigned To: {task.assignedTo}
-                        </p>
-                      </div>
-                    </div>
+                    <TaskCard key={task.id} task={task} onEdit={() => { setSelectedTask(task); setIsEditOpen(true); }} />
                   ))
               )}
             </div>
           </div>
         ))}
       </div>
+
+      {selectedTask && (
+        <EditTaskForm task={selectedTask} isOpen={isEditOpen} onClose={() => setIsEditOpen(false)} closeModal={function (): void {
+          throw new Error("Function not implemented.");
+        } } />
+      )}
     </div>
   );
 };
