+++ frontend/app/features/kanban/hooks/useEditTaskForm.ts
@@ -3,7 +3,7 @@ import { useEditTaskMutation } from "@store/api/taskSlice";
 import { useState } from "react";
 import { useForm } from "react-hook-form";
 import { TaskSchema, taskSchema } from "../schemas/taskSchema";
-import { Task } from "../utils/filterTasks";
+import { Task } from "../utils/Tasks";
 
 export function useEditTaskForm(task: Task, onClose: () => void) {
   const [isOpen, setIsOpen] = useState(false);
@@ -18,12 +18,12 @@ export function useEditTaskForm(task: Task, onClose: () => void) {
     resolver: zodResolver(taskSchema),
     defaultValues: {
       title: task.title,
-      taskOrder: task.taskOrder,
+      taskOrder: Number(task.taskOrder),
       description: task.description,
-      assignedTo: task.assignedTo ? Number(task.assignedTo) : 0,
+      assignedTo: task.assignedTo ? Number(task.assignedTo) : undefined,
       status: task.status || "TODO", 
       dueDate: task.dueDate ? task.dueDate.split("T")[0] : "",
-    },
+    }    
   });
 
   const onSubmit = async (data: TaskSchema) => {
