+++ frontend/app/features/kanban/utils/Tasks.ts
@@ -1,0 +1,11 @@
+export interface Task {
+  taskOrder: number;
+  id: number;
+  title: string;
+  description?: string;
+  status: string;
+  createdBy: number;
+  assignedTo?: number | null;
+  dueDate?: string | null;
+  createdAt: string;
+}
